use new lists for Yes/No markets, organized templates to get a step ahead toward market validation
