+++ packages/augur-ui/src/modules/types.ts
@@ -14,7 +14,7 @@ import { EthersSigner } from 'contract-dependencies-ethers/build/ContractDepende
 import { Getters, PayoutNumeratorValue } from '@augurproject/sdk';
 import { TransactionMetadataParams } from 'contract-dependencies-ethers/build';
 import { BigNumber } from 'utils/create-big-number';
-import { Template } from 'modules/create-market/get-template';
+import { TemplateInputType, REQUIRED, CHOICE } from 'modules/create-market/constants';
 
 export enum SizeTypes {
   SMALL = 'small',
@@ -731,4 +731,89 @@ export interface SortedGroup {
   label: string;
   subGroup?: Array<SortedGroup>;
   autoCompleteList?: Array<SortedGroup>;
+}
+
+
+export interface UserInputText {
+  value: string;
+}
+
+export interface UserInputDateYear extends UserInputText {}
+export interface UserInputDateTime {
+  endTime: number;
+  hour?: number;
+  minute?: number;
+  meridiem: string;
+  timezone: string;
+  offset: number;
+  offsetName: string;
+  endTimeFormatted: TimezoneDateObject;
+}
+export interface UserInputDropdown extends UserInputText {}
+export interface UserInputUserOutcome extends UserInputText {}
+
+export type UserInputtedType =
+  | UserInputText
+  | UserInputDateYear
+  | UserInputDateTime
+  | UserInputDropdown
+  | UserInputUserOutcome;
+
+export interface CategoryList {
+  [category: string]: [
+    {
+      [category: string]: [
+        {
+          [index: number]: string;
+        }
+      ];
+    }
+  ];
+}
+export interface TemplateChildren {
+  [category: string]: CategoryTemplate;
+}
+
+export interface CategoryTemplate {
+  templates: Template[];
+  children: TemplateChildren;
+}
+
+export interface ResolutionRule {
+  text: string;
+  isSelected?: boolean;
+}
+
+export interface ResolutionRules {
+  [REQUIRED]?: ResolutionRule[];
+  [CHOICE]?: ResolutionRule[];
+}
+
+export interface Template {
+  categories: Categories;
+  marketType: string;
+  question: string;
+  example: string;
+  inputs: TemplateInput[];
+  resolutionRules: ResolutionRules;
+  denomination?: string;
+  tickSize?: number;
+}
+
+export interface TemplateInput {
+  id: number;
+  type: TemplateInputType;
+  placeholder: string;
+  label?: string;
+  tooltip?: string;
+  userInput?: string;
+  userInputObject?: UserInputtedType;
+  values?: ValueLabelPair[];
+  sublabel?: string;
+}
+
+export interface Categories {
+  primary: string;
+  secondary: string;
+  tertiary: string;
 }
